Cleaning and adding entrypoints
