--- preprocessing/data.py
@@ -1,34 +1,0 @@
-from __future__ import division, absolute_import, print_function
-import numpy as np
-import pandas as pd
-
-
-def load_dataset(*, file_name: str) -> pd.DataFrame:
-    """
-    load csv dataset from path
-    :return: (df) pandas dataframe
-    """
-    # _data = pd.read_csv(f"{config.DATASET_DIR}/{file_name}")
-    _data = pd.read_csv(file_name)
-    return _data
-
-
-def data_split(df, start, end):
-    """
-    split the dataset into training or testing using date
-    :param data: (df) pandas dataframe, start, end
-    :return: (df) pandas dataframe
-    """
-    data = df[(df.date >= start) & (df.date < end)]
-    return format_for_env(data)
-
-
-def format_for_env(df):
-    df = df.sort_values(["date", "tic"], ignore_index=True)
-    df.index = df.date.factorize()[0]
-    return df
-
-def convert_to_datetime(time):
-    time_fmt = "%Y-%m-%dT%H:%M:%S"
-    if isinstance(time, str):
-        return datetime.datetime.strptime(time, time_fmt)
